+++ changelog.md
@@ -6,6 +6,7 @@
 - Added hardcore gamer mode on the Crimson Isle Tab
 **pre2:**
 - Fixed hardcore gamer mode not triggering
+- Attribute rendering now renders on Magma Lord Gauntlets
 # v0.3.2
 - Added saving before the update happens so you don't loose any data
 - Disabling party commands or party warp now also disables custom warp as it can break other mods.