+++ changelog.md
@@ -12,6 +12,9 @@
 - Added !invite name name2 name3 name4 dm command, which prompts the other player to accept and invite all those people.
 - Added a 100ms delay before sending the [Click to party] message so it always appears after the message
 - Added a !merge username command, which should attempt to merge the party.
+**pre3**
+- Removed check for the OK! thingy from merging (nerdy stuff)
+- Made !merge command require leader to trigger
 # v0.3.2
 - Added saving before the update happens so you don't loose any data
 - Disabling party commands or party warp now also disables custom warp as it can break other mods.